Automatically compile C library on install
This automatically compiles km_dict.cpp on install using setuptools'
Extension functionality. Strictly speaking, we are not really
compiling a true extension since we interface with the code using
ctypes. This is also why we need the hacky setup.py code to remove
PyInit_km_dict from the list of exported symbols. If this is not
done, we will get link errors on Windows. A better solution must
exist, but this works for now (I think).