Enable multi-image training for gauss feat. insegt
This was a "shortest path" update, so it may not have been the
optimal way of doing it. However, while not fully tested, it does
seem to work. Main bottleneck is that updating the label dict.
also requires that we label all the training images - this takes
a lot of time and a lot of memory. Need a more thorough rewrite to
remove that.